+++ azext_iot/operations/central.py
@@ -37,8 +37,9 @@ def iot_central_monitor_events(cmd, app_id, device_id=None, consumer_group='$Def
     import importlib
 
     events3 = importlib.import_module('azext_iot.operations.events3._events')
+    EventTargetBuilder = importlib.import_module('azext_iot.operations.events3.EventTargetBuilder')
 
-    eventHubTarget = events3.buildCentralEventHubTargetSync(cmd, app_id, aad_token)
+    eventHubTarget = EventTargetBuilder.buildCentralEventHubTargetSync(cmd, app_id, aad_token)
     events3.executor(eventHubTarget,
                      consumer_group=consumer_group,
                      enqueued_time=enqueued_time,
