+++ azext_iot/operations/events3/_events.py
@@ -71,81 +71,6 @@ def executor(target, consumer_group, enqueued_time, properties=None,
                 raise RuntimeError(error)
 
 
-def buildCentralEventHubTargetSync(cmd, app_id, aad_token):
-    eventLoop = asyncio.new_event_loop()
-    asyncio.set_event_loop(eventLoop)
-
-    centralTarget = eventLoop.run_until_complete(buildCentralEventHubTarget(cmd, app_id, aad_token))
-    return centralTarget
-
-
-async def buildCentralEventHubTarget(cmd, app_id, aad_token):
-    from azext_iot.common._azure import get_iot_central_tokens
-
-    tokens = get_iot_central_tokens(cmd, app_id, aad_token)
-    eventHubToken = tokens['eventhubSasToken']
-    hostnameWithoutPrefix = eventHubToken['hostname'].split("/")[2]
-    endpoint = hostnameWithoutPrefix
-    path = eventHubToken["entityPath"]
-    tokenExpiry = tokens['expiry']
-    auth = _build_auth_container_from_token(endpoint, path, eventHubToken['sasToken'], tokenExpiry)
-    address = "amqps://{}/{}/$management".format(hostnameWithoutPrefix, eventHubToken["entityPath"])
-    meta_data = await query_meta_data(address, path, auth)
-    partition_count = meta_data[b'partition_count']
-    partition_ids = []
-    for i in range(int(partition_count)):
-        partition_ids.append(str(i))
-    partitions = partition_ids
-    auth = _build_auth_container_from_token(endpoint, path, eventHubToken['sasToken'], tokenExpiry)
-
-    eventHubTarget = {
-        'endpoint': endpoint,
-        'path': path,
-        'auth': auth,
-        'partitions': partitions
-    }
-
-    return eventHubTarget
-
-
-def buildIotHubTargetSync(target):
-    eventLoop = asyncio.new_event_loop()
-    asyncio.set_event_loop(eventLoop)
-
-    hubTarget = eventLoop.run_until_complete(buildIotHubTarget(target))
-    return hubTarget
-
-
-async def buildIotHubTarget(target):
-    if 'events' not in target:
-        endpoint = _build_iothub_amqp_endpoint_from_target(target)
-        _, update = await evaluate_redirect(endpoint)
-        target['events'] = update['events']
-        endpoint = target['events']['endpoint']
-        path = target['events']['path']
-        auth = _build_auth_container(target)
-        meta_data = await query_meta_data(target['events']['address'], target['events']['path'], auth)
-        partition_count = meta_data[b'partition_count']
-        partition_ids = []
-        for i in range(int(partition_count)):
-            partition_ids.append(str(i))
-        target['events']['partition_ids'] = partition_ids
-    else:
-        endpoint = target['events']['endpoint']
-        path = target['events']['path']
-    partitions = target['events']['partition_ids']
-    auth = _build_auth_container(target)
-
-    iotHubTarget = {
-        'endpoint': endpoint,
-        'path': path,
-        'auth': auth,
-        'partitions': partitions
-    }
-
-    return iotHubTarget
-
-
 async def initiate_event_monitor(target, consumer_group, enqueued_time, device_id=None, properties=None,
                                  timeout=0, output=None, content_type=None, devices=None, interface_id=None, pnp_context=None):
     def _get_conn_props():
@@ -298,56 +223,6 @@ async def monitor_events(endpoint, connection, path, auth, partition, consumer_g
         if not exp_cancelled:
             await receive_client.close_async()
         logger.info("Closed monitor on partition %s", partition)
-
-
-def _build_auth_container(target):
-    sas_uri = 'sb://{}/{}'.format(target['events']['endpoint'], target['events']['path'])
-    return uamqp.authentication.SASTokenAsync.from_shared_access_key(sas_uri, target['policy'], target['primarykey'])
-
-
-def _build_auth_container_from_token(endpoint, path, token, tokenExpiry):
-    sas_uri = 'sb://{}/{}'.format(endpoint, path)
-    return uamqp.authentication.SASTokenAsync(audience=sas_uri, uri=sas_uri, expires_at=tokenExpiry, token=token)
-
-
-async def evaluate_redirect(endpoint):
-    source = uamqp.address.Source('amqps://{}/messages/events/$management'.format(endpoint))
-    receive_client = uamqp.ReceiveClientAsync(source, timeout=30000, prefetch=1, debug=DEBUG)
-
-    try:
-        await receive_client.open_async()
-        await receive_client.receive_message_batch_async(max_batch_size=1)
-    except uamqp.errors.LinkRedirect as redirect:
-        redirect = unicode_binary_map(parse_entity(redirect))
-        result = {}
-        result['events'] = {}
-        result['events']['endpoint'] = redirect['hostname']
-        result['events']['path'] = redirect['address'].replace('amqps://', '').split('/')[1]
-        result['events']['address'] = redirect['address']
-        return redirect, result
-    finally:
-        await receive_client.close_async()
-
-
-async def query_meta_data(endpoint, path, auth):
-    source = uamqp.address.Source(endpoint)
-    receive_client = uamqp.ReceiveClientAsync(source, auth=auth, timeout=30000, debug=DEBUG)
-    try:
-        await receive_client.open_async()
-        message = uamqp.Message(application_properties={'name': path})
-
-        response = await receive_client.mgmt_request_async(
-            message,
-            b'READ',
-            op_type=b'com.microsoft:eventhub',
-            status_code_field=b'status-code',
-            description_fields=b'status-description',
-            timeout=30000
-        )
-        test = response.get_data()
-        return test
-    finally:
-        await receive_client.close_async()
 
 
 def send_c2d_message(target, device_id, data, properties=None,
